requirements.txt
accept-paymob==1.0.0
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.2
billiard==4.1.0
boto3==1.28.20
botocore==1.31.20
celery==5.3.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
daphne==4.0.0
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-phonenumber-field==7.0.2
django-redis==5.2.0
django-silk==5.0.3
django-storages==1.13.2
django-tenants==3.4.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-writable-nested==0.7.0
drf-yasg==1.21.5
factory-boy==3.3.0
Faker==19.6.2
googletrans==3.1.0a0
gprof2dot==2022.7.29
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kafka-python==2.0.2
kombu==5.3.1
MarkupSafe==2.1.2
msgpack==1.0.5
openapi-codec==1.3.2
packaging==23.0
phonenumberslite==8.13.8
pika==1.3.2
Pillow==9.4.0
prompt-toolkit==3.0.39
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
service-identity==21.1.0
simplejson==3.18.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
stripe==5.2.0
tomli==2.0.1
txaio==23.1.1
typing-extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
zope.interface==5.5.2
